# spec/openapi-spec.yml
openapi: 3.0.3
info:
  title: In The Mealtime
  description: This is the API for In The Mealtime
  version: 1.0.0
...
paths:
...
  /api/restaurant:
    get:
      tags:
        - Restaurant
      summary: Fetch all restaurants
      description: Fetch all restaurants
      operationId: fetchRestaurants
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Restaurant'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Restaurant'
        400:
          $ref: "#/components/responses/ClientErrorResponse"
        500:
          $ref: "#/components/responses/ServerErrorResponse"
    post:
      tags:
        - Restaurant
      summary: Create a new restaurant
      description: Create a new restaurant
      operationId: createRestaurant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestaurantPatch'
          application/xml:
            schema:
              $ref: '#/components/schemas/RestaurantPatch'
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
            application/xml:
              schema:
                $ref: '#/components/schemas/Restaurant'
        400:
          $ref: "#/components/responses/ClientErrorResponse"
        500:
          $ref: "#/components/responses/ServerErrorResponse"
  /api/restaurant/{id}:
    get:
      tags:
        - Restaurant
      summary: Fetch restaurant
      description: Fetch restaurant
      operationId: fetchRestaurant
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/ID"
      responses:
        200:
          $ref: "#/components/responses/SuccessRestaurantResponse"
        400:
          $ref: "#/components/responses/ClientErrorResponse"
        500:
          $ref: "#/components/responses/ServerErrorResponse"
    put:
      tags:
        - Restaurant
      summary: Updates a restaurants
      description: Updates a new restaurants
      operationId: updateRestaurant
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: header
          name: If-Match
          required: true
          schema:
            $ref: "#/components/schemas/Version"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestaurantPatch'
          application/xml:
            schema:
              $ref: '#/components/schemas/RestaurantPatch'
      responses:
        200:
          $ref: "#/components/responses/SuccessRestaurantResponse"
        400:
          $ref: "#/components/responses/ClientErrorResponse"
        '412':
          $ref: "#/components/responses/ClientErrorResponse"
        500:
          $ref: "#/components/responses/ServerErrorResponse"
    delete:
      tags:
        - Restaurant
      summary: Delete a restaurants
      description: Delete a new restaurants
      operationId: deleteRestaurant
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: header
          name: If-Match
          required: true
          schema:
            $ref: "#/components/schemas/Version"
      responses:
        204:
          description: Successful operation
        400:
          $ref: "#/components/responses/ClientErrorResponse"
        412:
          $ref: "#/components/responses/ClientErrorResponse"
        500:
          $ref: "#/components/responses/ServerErrorResponse"
  /api/restaurant/{id}/report:
    get:
      tags:
        - Restaurant
      summary: Fetch restaurant report
      description: Fetch restaurant report
      operationId: fetchRestaurantReport
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/ID"
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantReport"
            application/xml:
              schema:
                $ref: "#/components/schemas/RestaurantReport"
        400:
          $ref: "#/components/responses/ClientErrorResponse"
        500:
          $ref: "#/components/responses/ServerErrorResponse"
  /api/restaurant/{restaurantId}/menu-page:
    put:
      tags:
        - Restaurant
      summary: Creates a new menu page of a specified restaurant
      description: Creates a new menu page of a specified restaurant
      operationId: addRestaurantsMenuPage
      parameters:
        - in: path
          name: restaurantId
          required: true
          schema:
            $ref: "#/components/schemas/ID"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
            encoding:
              file:
                contentType: image/png, image/jpeg, image/jpg, image/gif, image/webp
                headers:
                  Content-Type:
                    schema:
                      type: string
                      enum: [ image/png, image/jpeg, image/jpg, image/gif, image/webp ]
      responses:
        200:
          $ref: "#/components/responses/SuccessRestaurantResponse"
        400:
          $ref: "#/components/responses/ClientErrorResponse"
        500:
          $ref: "#/components/responses/ServerErrorResponse"
  /api/restaurant/{restaurantId}/menu-page/{pageId}:
    get:
      tags:
        - Restaurant
      summary: Fetch menu page of a specified restaurant
      description: Fetch menu page of a specified restaurant
      operationId: fetchRestaurantsMenuPage
      parameters:
        - in: path
          name: restaurantId
          required: true
          schema:
            $ref: "#/components/schemas/ID"
        - in: path
          name: pageId
          required: true
          schema:
            $ref: "#/components/schemas/ID"
        - in: query
          name: thumbnail
          required: false
          schema:
            type: boolean
      responses:
        200:
          description: Successful operation
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/gif:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
        400:
          $ref: "#/components/responses/ClientErrorResponse"
        500:
          $ref: "#/components/responses/ServerErrorResponse"
    delete:
      tags:
        - Restaurant
      summary: Delete a menu page of a specified restaurant
      description: Delete a menu page of a specified restaurant
      operationId: deleteRestaurantsMenuPage
      parameters:
        - in: path
          name: restaurantId
          required: true
          schema:
            $ref: "#/components/schemas/ID"
        - in: path
          name: pageId
          required: true
          schema:
            $ref: "#/components/schemas/ID"
      responses:
        200:
          $ref: "#/components/responses/SuccessRestaurantResponse"
        400:
          $ref: "#/components/responses/ClientErrorResponse"
        500:
          $ref: "#/components/responses/ServerErrorResponse"
components:
  schemas:
    RestaurantPatch:
      type: object
      properties:
        name:
          type: string
        style:
          type: string
        kind:
          type: string
        phone:
          type: string
        website:
          type: string
        email:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        address:
          $ref: '#/components/schemas/Address'
    Restaurant:
      type: object
      required:
        - id
        - name
        - address
        - menuPages
      allOf:
        - $ref: "#/components/schemas/RestaurantPatch"
        - $ref: "#/components/schemas/Updatable"
      properties:
        id:
          $ref: "#/components/schemas/ID"
        address:
          $ref: '#/components/schemas/Address'
        menuPages:
          type: array
          items:
            $ref: "#/components/schemas/MenuPage"
    Address:
      type: object
      properties:
        street:
          type: string
        housenumber:
          type: string
        postal:
          type: string
        city:
          type: string
    RestaurantReport:
      type: object
      required:
        - restaurantId
        - countOfOrders
        - countOfOrderedMeals
        - overallPrice
        - overallTip
        - topFetchers
        - topOrderers
        - topMoneyCollectors
        - topParticipants
        - topMeals
      properties:
        restaurantId:
          $ref: "#/components/schemas/ID"
        countOfOrders:
          type: integer
          format: int32
        countOfOrderedMeals:
          type: integer
          format: int32
        overallPrice:
          type: number
          format: float
        overallTip:
          type: number
          format: float
        topOrderers:
          type: array
          items:
            $ref: "#/components/schemas/StatisticPerson"
        topFetchers:
          type: array
          items:
            $ref: "#/components/schemas/StatisticPerson"
        topMoneyCollectors:
          type: array
          items:
            $ref: "#/components/schemas/StatisticPerson"
        topParticipants:
          type: array
          items:
            $ref: "#/components/schemas/StatisticPerson"
        topMeals:
          type: array
          items:
            $ref: "#/components/schemas/StatisticPerson"
    StatisticPerson:
      type: object
      properties:
        name:
          type: string
        count:
          type: integer
          format: int32
    MenuPage:
      type: object
      required:
        - id
        - name
        - index
      properties:
        id:
          type: string
          format: uuid
        index:
          type: integer
          format: int32
        name:
          type: string
...
    ID:
      type: string
      format: uuid
    Version:
      type: string
      format: uuid
    ErrorObject:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        correlationId:
          type: string
        message:
          type: string
        detailMessage:
          type: string
        stacktrace:
          type: string
    Creatable:
      type: object
      required:
        - createdAt
        - createdBy
      properties:
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
    Updatable:
      type: object
      required:
        - version
      allOf:
        - $ref: "#/components/schemas/Creatable"
      properties:
        version:
          $ref: "#/components/schemas/Version"
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
...
  responses:
...
    SuccessRestaurantResponse:
      description: Successful operation
      headers:
        ETag:
          schema:
            $ref: "#/components/schemas/Version"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Restaurant'
        application/xml:
          schema:
            $ref: '#/components/schemas/Restaurant'
    ClientErrorResponse:
      description: Operation failed due to client error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorObject"
        application/xml:
          schema:
            $ref: "#/components/schemas/ErrorObject"
    ServerErrorResponse:
      description: Operation failed due to client error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorObject"
        application/xml:
          schema:
            $ref: "#/components/schemas/ErrorObject"